Always output result of Opbeat exception logging
Previously a custom callback to `handleUncaughtExceptions` would
overload the normal handling of the call to Opbeat. Now this callback
acts more as a wrapper to the custom callback, which is then called
afterwards.

Notice, that if a custom callback is provided to
`handleUncaughtExceptions`, the process is NOT terminated. So if you use
this method with a custom callback, remember to terminate the
node-process your self.