Log invalid objects better
This is an attempt to help in the situation where someone tries to log
an 'invalid' object:

  opbeat.captureError({})

The client would previously log the error as 'undefined undefined', but
with this commit the actual object will be logged in a readable form.
This way it's easier to find the location in the source code where the
'invalid' object was generated.