New feature: Log location in code where error was discovered
This is currently behind a feature flag and needs to be enabled using
the boolean config option `_ff_captureFrame` or environment variable
`OPBEAT_FF_CAPTURE_FRAME`.

If enabled all regular errors (i.e. not uncaught exceptions) whos top
stack frame isn't `in_app` will have an extra stack frame added to the
top of the stack with the location where the call to `captureError` was
performed.