Recover lost or wrong transaction
Every time an Trace ends, we now check if the related transaction
matches the `currentTransaction` and overwrite it if that's not the
case.

This happens a lot with certain database modules, normally if they
implement their own connection pooling.

In that case db query callbacks are not queued on the event loop, but
added to an internal queue and called once the result comes back from
the database. This in turns means that the `currentTransaction` gets set
to whatever was the current transaction when the queing logic got
initialized - usually during the first request to the Node process.