mysql: improve grouping of mysql errors
To improve grouping of mysql errors that happens after the async
boundery, we modify the exception type to include the mysql error type.
This improves the grouping because the exception type is used as part of
the grouping whereas the error message is not.

This is related to https://github.com/opbeat/opbeat-node/issues/40