Add opbeat.addFilter()
The addFilter function expects a filter function as its only argument.
Each filter function will be called in the order they were added every
time an error is about to be logged with Opbeat.

Each filter function will receive a `data` object as the only argument
containing the data about to be sent to Opbeat.

If a filter function doesn't return any value or returns a falsy value,
the remaining filter functions will not be called and the error will not
be logged with Opbeat.

This commit deprecates the old `filter` option.