core: fix missing stack traces for certain traces
The transaction name should be used as part of the grouping key when
determining if a stack trace is already known (i.e. already sent to the
Opbeat instake API).

Since the transaction name is not always known in advance when starting
the trace, we now have to record the stack trace for all new traces in
case it's needed.

The only optimization we can do now is not trying to process it and send
it to the intake API if it's deemed to be known at flush time.