core: ensure transactions are properly ended in Node.js v0.10
In v0.10 we previously patched the `res.end` function to know when the
request transaction should be ended. This caused the transaction to be
ended prior to the transaction result and name being set (since that's
now set when `writeHead` is called).

Instead we now listen for the `finish` event as this appears to be
equivalent to the `prefinish` event in newer versions of Node.js.