knex: get better performance stack traces when using knex
On db traces we'd normally show a stack trace back to the users code where
a given query was made. When using knex it's not possible because all
queries handled by knex happens after an async boundary.

With this change we now overwrite that default stack trace with a custom
stack trace taken when the knex query builder was initiated (or a call
to `knex.raw` which doesn't use the knex query builder).