core: allow starting of Opbeat using `node -r` flag
Example usage:

```
node -r opbeat/start server.js
```

Opbeat now ships with a special script in the root of the package called
`start.js`. If you require this script instead of the regular main
script, Opbeat will be started automatically without the need for
calling the `start()` function.

By default Opbeat will take all its configuration from environment
variables - and this is also possible when using the new start script.
But as an added bonus, it's now possible to also specify all the Opbeat
configuration in an `opbeat.js` file that can be placed in the current
working directory. If this file is found, the start script will use that
to configer the Opbeat agent.

Example `opbeat.js` file:

```js
module.exports = {
  appId: '...',
  organizationId: '...',
  secretToken: '...'
}
```

Note that if you are using this feature, it's no longer required to
require the `opbeat` module within your application and you should NOT
call the `start()` function manually.