core: allow setting per-transaction context data
This adds two new functions to the Opbeat agent:

- `opbeat.setUserContext(context)`
- `opbeat.setExtraContext(context)`

The given `context` argument must be an object.

The given `context` will be added to the active transaction. If no
active transaction can be found, `false` is returned. Otherwise `true`.

It's possible to call these functions multiple times within the scope of
the same active transaction. In that case the properties of the
`context` object will be shallow merged with context previously given.

It's also possible to call these functions directly on a transaction
object. But usually you shouldn't need to do that. These functions will
not return any boolean.

The provided context is displayed along side other contextual
information under the Performance section on opbeat.com.

If an error is captured, the context from the active transaction is used
as context for the captured error and any custom context given as the
2nd argument to `opbeat.captureError` is shallow merged ontop.

For more details on contextual information, see:
https://opbeat.com/blog/posts/introducing-contextual-information-in-performance-transactions/