core: add flushInterval config option
The Opbeat agent maintains an in-memory queue to which recorded metrics
and sampled transactions are added when they end. By default, this queue
is flushed and sent to Opbeat for processing every 60 seconds.

Use this option to change that interval. The value is expected to be in
seconds. The environment variable alternative is called
`OPBEAT_FLUSH_INTERVAL`.

Lowering this interval can reduce memory usage on Node.js applications
with a high number of transactions.