core: add source map support
Opbeat will detect the source map by looking for the magic
`sourceMappingURL` comment in your transpiled source code.

Both inline and linked source mapped are supported.

For optimal support, it's recommended that the original source code is
either embedded in the source map it self or discoverable on the
file-system using the source map.

Fixes #74