refactor(instrumentation): handle edge case where flushInterval < 5s
The purpose of the boot flush time is to be lower than the normal flush
time in order to get a result quickly when the app first boots. But if a
custom flush interval is provided and it's lower than the boot flush
time, it doesn't make much sense anymore. In that case, just pretend we
have already used the boot flush time.