refactor(http): log HTTP results in groups of 100 (#68)
When recording an HTTP transaction we set its result to the HTTP status
code. With this change we now only record the major group, e.g. 2xx
instead of 204 and 5xx instead of 500 etc.

The transaction name will also be prefixed with "HTTP " to more clearly
indicate that it's an HTTP transaction.