test: fix flaky `disableInstrumentations` tests (#541)
Sometimes the `test/config.js` tests would run long enough for one of
the agents created during the tests to flush its queue. The agent had
its filters mocked and wasn't expected to actualy send data anywhere, so
this would result in an uncaught exception.

This fix does away with mocking of the filters as there was really no
reason for it. It also patches the test that wrongly added a transaction
to the queue. So now the queue should never be flushed no matter how
long the tests run.

Fixes #535