test(ws): fix tests for ws on older Node versions (#719)
In preparation of us starting to backport to the 1.x branch, we need to make sure the tests pass first. Recently a new version of `ws` was released that wasn't compatible with Node.js 4 which made our tests fail. This PR fixes that by omitting it from our TAV tests.