feat(span): add tags to spans (#757)
Add tag support to spans by adding the following two functions to the
Span object:

- `span.setTag()`
- `span.addTags()`

As a side-effect of the refactor, the following TraceContext related
API's will also be exposed on the Span object:

- `span.parentId`
- `span.traceId`
- `span.sampled`

Since version 2.0.0, the TraceContext was exposed on spans and
transactions via a `context` property. This was never documented, and
this commit renames the property to `_context`.