docs: add info about when API's were added (#1000)
I left out a few API's as those actually need to list dates when they changed (e.g. the `agent.addFilter()` was added in v0.1.0, but the API changed in v2.0.0). But I didn't want to tackle that in this PR.

I also touched up the new `traceparent` API's to use the new docs layout.