chore: prepare code style for new version of linter (#1284)
We can't upgrade to the new version of 'standard' currently because we
still use the deprecated `url.parse` API in several locations (see issue
no #1254). But to ensure fewer merge conflicts between the 2.x and the
upcoming 3.x branch, we'll fix as many future linting issues as
possible already now.