fix: avoid import of any '@types/...' packages in types file (#2335)
This ensures that TypeScript users do not need to install separate
'@types/...' packages to build -- with the lone exception of
'@types/node'.

We avoid imports by instead inlining the required types. This is a
trade-off. We don't want '@types/...' in "dependencies" because they
can be big and aren't really "official" types for those packages.

Fixes: #2331