tests: fix TAV test breakage with node 17 (hapi <=16, fastify <0.36.0) (#2387)
This was broken in #2380. Versions of hapi <=16 did not support the
'host' option to `new Server`, but instead to `server.connection`.
Also really old versions of fastify (v0.35.7) cannot even be npm installed
with the npm included in node v17. This drops support for fastify <1.0.0
and updates the TAV test matrix to only test a given fastify version
with its supported versions of node per https://www.fastify.io/docs/latest/LTS/