feat: zero-configuration support by always successfully inferring a serviceName (#2410)
If a bogus serviceName is explicitly configured, the agent will still
log.error and de-activate. However, if the agent is inferring a
serviceName, then with this change it will always succeed. This change
improves inference from a package.json by dealing with some cases where
a package.json "name" was an invalid serviceName. And then ultimately it
falls back to "nodejs_service" if a name cannot otherwise be inferred.

This changes Agent configuration somewhat. Before this change the agent
would `this._config()` twice: once at Agent creation (resulting in a
agent._conf that is incorrect if options are later passed to
agent.start()), and once at agent.start(). This commit changes to only
configure at agent.start(). A separate `config.configLogger()` handles
setting `agent.logger` to a valid early logger when the Agent is
created.  All Agent methods have been updated as necessary to guard
against usage before the agent is started.  One visible semantic change
is that `agent.getServiceName()` will now always return `undefined`
before the agent is started. Arguably the earlier behaviour was a
subtle bug, but the returned serviceName could be *wrong* if the agent
was later started with a given serviceName.

Closes: #1944