fix: offline-queued redis commands would be double-instrumented (#2446)
Ensure that we do not create a second span for offline-queued redis
commands (those commands that redis queues before the client is
"ready").

Also add tests for and fix (probably unnecessarily) the instrumentation
breaking the calling form `client.cmd([args..., myCb])` where the
callback is in a first arg array. This form was supported in redis <=2.4.2
(i.e. only redis client releases from 2015 and earlier :).