feat: exit spans, infer destination.service.resource
This adds the explicit concept of "exit spans" -- spans the represent an
outgoing call to a downstream service. An exit span doesn't have child
spans. (Technically it can have spans of the same type+subtype, but this
agent doesn't have any cases of that.) An exit span is explicitly marked
at creation time by setting the `exitSpan` option to true.
https://github.com/elastic/apm/blob/master/specs/agents/tracing-spans.md#exit-spans

For exit spans the 'span.context.destination.service.resource' is
automatically inferred from other context. This can be overriden with
the new `span.setDestinationService(resource)`.
https://github.com/elastic/apm/blob/master/specs/agents/tracing-spans-destination.md

Fixes: #2103