fix: ensure the span run-context for outgoing http requests does not spill into user code (#2470)
This fixes the 'http' and 'https' instrumentation for outgoing requests
to not have the 'http' span context be active in user code. This ensures
that user code cannot create a child span of the http span, which would
(a) be misleading and (b) cause problems for coming exit span and
compressed span work.

Also, fix a bug in the https instrumentation in older versions of node
(version <9.0.0) where the instrumentation of `https.request` relied on
intercepting `http.request` (that Node's `https.request()` would call). 
The agent didn't guarantee that the 'http' module was instrumented. 
A user program that used `https.request` without indirectly
`require('http')`ing would not get an HTTP span.

Refs: #2430