fix: ensure correct run context for '@elastic/elasticsearch' instrumentation (#2550)
This is a re-write of the @elastic/elasticsearch instrumentation that
stops using the ES client observability events
    https://www.elastic.co/guide/en/elasticsearch/client/javascript-api/current/observability.html
and switches to patching Transport#request() instead. This is necessary
to be able to bind that `request()` invocation to the RunContext for the
Span we've created, without using `apm.startSpan(...)`, which bleeds the
RunContext out to user code. (Patching Transport#request() is what the
legacy 'elasticsearch' instrumentation is also doing.)

Also: '@elastic/elasticsearch-canary' TAV tests were failing because an
ES service container was not being setup for its test run; we didn't see
this in CI before because the Jenkins TAV config doesn't have the
canary module.

Refs: #2430