chore: switch from standard to eslint (#2579)
This switches from using standard to eslint for linting/formatting.
It is currently setup to use the same "standard" rules. This change
allows us to add custom rules, should we want to, or eventually to
diverge on style.

Using "ignorePatterns" in .eslintrc.json has some duplication of
.gitignore, but is necessary for some ignoring of fixture files that
are in git.