test: attempt to avoid spurious failure over temporarily down httpstat.us (#2654)
There were a number of spurious test failures in http-run-context.test.js
earlier today during a relatively short time window. All were when
making a request to httpstat.us. A theory is httpstat.us was down
(perhaps for maint) for that time window. Let's switch to using
www.google.com for test requests instead (as was done earlier in #2521
for a similar reason.