test: revert 'docker inspect ...' dump, it makes test failure logs unreadable (#2685)
This reverts part of the change from #2678. The intent had been to dump
some info to help with "Container '...' is unhealthy" errors.  However,
the main side-effect is that the end of the log file is 1000s of docker
inspect lines and a test failure is nowhere near the end of the log
file.