feat: add span links for any 'traceparent' attrs in SQS/SNS-triggered Lambdas (#2708)
For SQS- or SNS-triggered Lambda invocations, if any of the message
records include trace-context (i.e. a 'traceparent' entry in message
attributes), then a span link will be created to that traceparent on the
transaction for the Lambda invocation. This is the second part of #2593.

This also updates the collected metadata for SQS- and SNS-triggered
Lambda invocations per the latest spec: the special casing of
invocations with a *single* record has been reverted.

Refs: https://github.com/elastic/apm/issues/605
Closes: #2593