chore: include NOTICE file in npm published package and in dist (#2764)
The NOTICE.md file in the distribution includes the license for all
the bundled packages (i.e. add dependencies, excluding devDependencies).
Creating the distribution (dev-utils/make-distribution.sh) will fail
if one of the deps has no license or an unknown license.

Also add 'npm run lint:license-files' so that lint fails if a new dep doesn't
have a license or license file that will be needed for release.