feat: replace metrics impl with one based on @otel/sdk-metrics-base
OpenTelemetry JS Metrics is nearing 1.0. (The Java, Python, and Go
OTel projects have, I think, reached 1.0.) This SDK might be a good
replacement for the node-measured libs we are using for handling
metrics. It could be a path to working on current limitations in our
metrics support.

This commit is a first stab at trying out @otel/sdk-metrics-base for our
metrics generation. Supported so far:
- The `apm.registerMetric()` only public API for adding custom metrics.
- Runtime metrics (`nodejs.*`).

Not yet:
- breakdown metrics
- system mertics
- queue metrics for SQS instrumentation
- perf comparison
- etc.