fix: lambda instrumentation must set transaction *default* name to allow subsequent override by graphql instrumentation (#2865)
Before this the Lambda instrumentation was setting `trans.name = ...`,
which (via the Transaction.name setter) would set `trans._customName`,
which has higher prio than the `trans._defaultName`. Setting
`_customName` is intended for the *user* to override. If a Lambda
function is using GraphQL, the graphql instrumentation should be able to
override the transaction name.

Fixes: #2832