setDestinationContext
This adds instrumentation of RedisClient.create (used for the public
`redis.createClient(options)`) to pass the client `options` down to the
instrumentation of RedisCommandsQueue that is creating the redis spans.
The client options are stored on the `Instrumentation` singleton for the
duration of the `RedisClient.create`. Because the creation of the
client's RedisCommandsQueue is done synchronously in its constructor,
this works to unambiguously pass the options for that particular
client/queue.